feat: UnAuthenticatedException 발생시 error 페이지 이동
1. advice 이름 변경
   - restControllerAdvice: Api~~ExceptionHandler
   - ControllerAdvice: ~~ExceptionHandler
2. error.html 파일 생성
   - 로그인/회원가입 modal만 존재하는 페이지
   - 해당 modal을 닫으면, 메인 페이지로 이동
