+++ src/main/java/com/kvp/thinkjeonju/controller/MyPageController.java
@@ -1,0 +1,4 @@
+package com.kvp.thinkjeonju.controller;
+
+public class MyPageController {
+}
