+++ src/main/resources/templates/error.html
@@ -1,0 +1,86 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>think_jeonju</title>
+    <link href="https://fonts.googleapis.com/css?family=Noto+Sans+KR:300,400,500,700&amp;subset=korean" rel="stylesheet">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
+
+    <link rel="stylesheet" href="/css/base.css"/>
+</head>
+<body>
+<!-- Login Modal -->
+<div class="modal fade" id="login-modal" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
+    <div class="modal-dialog modal-dialog-centered" role="document">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="login-modal-title"> 로그인 </h5>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body">
+                <form id="login-form">
+                    <div class="modal-input-area">
+                        <span>아이디</span>
+                        <input class="id" id="login-id" name="id"/>
+                        <div class="validation-error" id="login-id-error" style="display: none"> <span></span></div>
+                    </div>
+                    <div class="modal-input-area">
+                        <span>비밀번호</span>
+                        <input type="password" class="password" id="login-password" name="password"/>
+                    </div>
+                    <div class="validation-error" id="login-password-error" style="display: none"> <span></span></div>
+
+                    <div class="validation-error" id="login-error" style="display: none"> <span></span></div>
+                    <button type="button" class="btn btn-primary width100p" id="login-btn">로그인</button>
+                </form>
+                <button type="button" class="btn btn-outline-secondary width100p" onclick="" id="join-form-btn">회원가입</button>
+            </div>
+        </div>
+    </div>
+</div>
+<!-- Join Modal -->
+<div class="modal fade" id="join-modal" tabindex="-1" role="dialog" aria-labelledby="exampleModalCenterTitle" aria-hidden="true">
+    <div class="modal-dialog modal-dialog-centered" role="document">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h5 class="modal-title" id="join-modal-title"> 회원가입 </h5>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                    <span aria-hidden="true">&times;</span>
+                </button>
+            </div>
+            <div class="modal-body">
+                <form id="join-form">
+                    <div class="modal-input-area">
+                        <span>아이디</span>
+                        <input class="id" id="join-id" name="id"/>
+                    </div>
+                    <div class="validation-error" id="join-id-error" style="display: none"> <span></span></div>
+                    <div class="modal-input-area">
+                        <span>비밀번호</span>
+                        <input type="password" class="password" id="join-password" name="password"/>
+                    </div>
+                    <div class="validation-error" id="join-password-error" style="display: none"> <span></span></div>
+                    <div class="modal-input-area">
+                        <span>이름</span>
+                        <input class="name" id="join-name" name="name"/>
+                    </div>
+                    <div class="validation-error" id="join-name-error" style="display: none"> <span></span></div>
+
+                    <div class="validation-error" id="join-error" style="display: none"> <span></span></div>
+                    <button type="button" class="btn btn-primary width100p" id="join-btn">회원가입</button>
+                </form>
+                <button type="button" class="btn btn-outline-secondary width100p" onclick="" id="login-form-btn">로그인</button>
+            </div>
+        </div>
+    </div>
+</div>
+<script src="/js/jquery-3.3.1.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
+<script src="/js/base.js"></script>
+<script src="/js/errorPage.js"></script>
+<script src="/js/login.js"></script>
+<script src="/js/join.js"></script>
+</body>
+</html>
